fix: cast EXTRACT(EPOCH) to float8 for oldest_pending_age_seconds
Postgres's EXTRACT(EPOCH FROM ...) returns NUMERIC, not double
precision; sqlx's f64 decode was silently failing every 5s refresh
cycle (logged as a warning, gauge never actually updated). Found
by actually running the demo script against a live process and
reading its logs, not by unit testing the query in isolation.

+++ crates/reliableq-api/src/metrics.rs
@@ -77,8 +77,12 @@ async fn refresh_once(pool: &PgPool) -> Result<(), sqlx::Error> {
             .set(count as f64);
     }
 
+    // Postgres's EXTRACT(EPOCH FROM ...) returns NUMERIC, not
+    // double precision, so an explicit cast is required — without it
+    // sqlx's f64 decode fails every refresh cycle (caught by actually
+    // running this against live postgres, not just compiling it).
     let oldest_pending_age_seconds: Option<f64> = sqlx::query_scalar(
-        "SELECT EXTRACT(EPOCH FROM (now() - created_at)) FROM jobs \
+        "SELECT EXTRACT(EPOCH FROM (now() - created_at))::float8 FROM jobs \
          WHERE status = 'PENDING' ORDER BY created_at ASC LIMIT 1",
     )
     .fetch_optional(pool)
